+++ Sources/Lazy/ResettableLazy.swift
@@ -20,11 +20,11 @@ public struct ResettableLazy<Value>: LazyContainer {
     
     /// Privatizes the inner-workings of this functional lazy container
     @ValueReference
-    private var guts: ResettableValueHolder<Value>
+    private var guts: ResettableValueHolder
     
     
     /// Allows other initializers to have a shared point of initialization
-    private init(_guts: ValueReference<ResettableValueHolder<Value>>) {
+    private init(_guts: ValueReference<ResettableValueHolder>) {
         self._guts = _guts
     }
     
@@ -78,6 +78,11 @@ public struct ResettableLazy<Value>: LazyContainer {
     
     /// Indicates whether the value has indeed been initialized
     public var isInitialized: Bool { _guts.wrappedValue.hasValue }
+    
+    
+    public mutating func initializeNow() {
+        guts.initializeNow()
+    }
     
     
     /// Resets this lazy structure back to its unset state. Next time a value is needed, it will be regenerated using
@@ -147,6 +152,18 @@ public enum LazyContainerResettableValueHolder<Value> {
         case .unset(initializer: _): return false
         }
     }
+    
+    
+    /// Immediately initializes the value in this holder.
+    ///
+    /// If this holder already contains a value, this does nothing
+    public mutating func initializeNow() {
+        switch self {
+        case .hasValue(value: _, initializer: _): return
+        case .unset(let initializer):
+            self = .hasValue(value: initializer(), initializer: initializer)
+        }
+    }
 }
 
 
@@ -156,7 +173,5 @@ public enum LazyContainerResettableValueHolder<Value> {
 public extension LazyContainer {
     
     /// Takes care of keeping track of the state, value, and initializer as needed
-    ///
-    /// - Attention: This will change in version 5, to be an alias to `LazyContainerResettableValueHolder<Value>`
-    typealias ResettableValueHolder = LazyContainerResettableValueHolder
+    typealias ResettableValueHolder = LazyContainerResettableValueHolder<Value>
 }